+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * new version Control.FullScreen.js
 * overviewmap: same icon options for all markers
 * overviewmap: target=_blank deleted
+* overviewmap: disable transients in backend
 
 ### 4.2.4 / 240423